Refactor e2e tests
Separate large diagraph_test file into separate
end to end tests.
