+++ packages/python/diagraph/__init__.py
@@ -1,4 +1,32 @@
 from .classes.diagraph import Diagraph as Diagraph
+from .classes.types import (
+    ErrorHandler as ErrorHandler,
+)
+from .classes.types import (
+    Fn as Fn,
+)
+from .classes.types import (
+    FunctionErrorHandler as FunctionErrorHandler,
+)
+from .classes.types import (
+    FunctionLogHandler as FunctionLogHandler,
+)
+from .classes.types import (
+    KeyIdentifier as KeyIdentifier,
+)
+from .classes.types import (
+    LogEventName as LogEventName,
+)
+from .classes.types import (
+    LogHandler as LogHandler,
+)
+from .classes.types import (
+    Rerunner as Rerunner,
+)
+from .classes.types import (
+    Result as Result,
+)
 from .decorators.prompt import prompt as prompt
+from .llm.llm import LLM as LLM
 from .llm.openai_llm import OpenAI as OpenAI
 from .utils.depends import Depends as Depends
